+++ src/main/java/roomescape/application/auth/dto/TokenRequest.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.application.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
