+++ src/main/java/roomescape/application/auth/dto/TokenResponse.java
@@ -1,4 +1,4 @@
-package roomescape.auth.dto;
+package roomescape.application.auth.dto;
 
 public record TokenResponse(String accessToken) {
 }
