+++ src/main/java/roomescape/presentation/controller/ReservationController.java
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.application.ReservationService;
+import roomescape.application.auth.dto.MemberAuthRequest;
 import roomescape.application.dto.ReservationDto;
 import roomescape.application.dto.UserReservationCreateDto;
-import roomescape.auth.dto.MemberAuthRequest;
 import roomescape.infrastructure.AuthenticationPrincipal;
 
 @RestController
