+++ src/test/java/roomescape/infrastructure/JdbcThemeRepositoryTest.java
@@ -13,6 +13,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.ActiveProfiles;
 import roomescape.domain.Theme;
+import roomescape.infrastructure.jdbc.JdbcThemeRepository;
 
 @JdbcTest
 @Import(JdbcThemeRepository.class)
